requirements.txt
fastapi
uvicorn
isort
autoimport
autopep8
flake8
pylint
pydantic
psycopg2
sqlalchemy
pydantic[dotenv]
email-validator
passlib[bcrypt]
python-jose[cryptography]
pyopenssl
oauth2client
python-multipart
